extended avatar component and merged to one file
