+++ client/src/components/Avatar.js
@@ -1,17 +1,37 @@
 import React from 'react';
 import styled from '@emotion/styled';
+import cameraIcon from '../assets/camera.svg';
 import PlaceholderAvatar from '../assets/profile/avatarPlaceholder.svg';
-import PropTypes from 'prop-types';
 
 const ProfileImg = styled.img`
-  min-height: 100px;
+  height: 100px;
   width: 100px;
   border-radius: 50%;
-  background: url(${PlaceholderAvatar}) no-repeat;
   background-size: cover;
   border: 4px solid #8645ff;
   object-fit: cover;
   background-color: white;
+`;
+
+const UploadImage = styled.button`
+  display: flex;
+  height: 25px;
+  width: 25px;
+  margin-left: 35px;
+  position: absolute;
+  align-items: center;
+  justify-content: center;
+  background: url(${cameraIcon}) no-repeat;
+  background-position: center;
+  border-radius: 50%;
+  border: none;
+  background-color: white;
+  outline: none;
+  cursor: pointer;
+  &:active {
+    transform: translateY(1px) scale(0.9);
+    box-shadow: 0px 5px 10px rgba(0, 0, 0, 20%);
+  }
 `;
 
 const ImageInput = styled.input`
@@ -35,22 +55,46 @@ const AvatarContainer = styled.div`
   justify-content: center;
   align-items: flex-end;
 `;
-const Avatar = (onChange) => {
+const Avatar = () => {
+  const [image, setImage] = React.useState(null);
+  const [upload, setUpload] = React.useState(true);
+
+  const handleImageUpload = async (event) => {
+    event.preventDefault();
+    const files = event.target.files;
+    const formData = new FormData();
+    formData.append('file', files[0]);
+    formData.append('upload_preset', 'bouybyw8');
+    setUpload(false);
+    const uploadedImage = await fetch(
+      'https://api.cloudinary.com/v1_1/dqbs4ljft/image/upload',
+      {
+        method: 'POST',
+        body: formData,
+      }
+    );
+    const file = await uploadedImage.json();
+    setImage(file.secure_url);
+  };
+
   return (
-    <>
-      <AvatarContainer>
-        <ImageInput
-          type="file"
-          onChange={onChange}
-          accept=".png, .jpg, .jpeg"
-        />
-        <ProfileImg src="https://res.cloudinary.com/dqbs4ljft/image/upload/v1588503708/uyaqo55cn1brdosoiwgu.jpg" />
-      </AvatarContainer>
-    </>
+    <AvatarContainer>
+      {upload ? (
+        <>
+          <ProfileImg src={PlaceholderAvatar} />
+          <UploadImage>
+            <ImageInput
+              type="file"
+              onChange={handleImageUpload}
+              accept=".png, .jpg, .jpeg"
+            />
+          </UploadImage>
+        </>
+      ) : (
+        <ProfileImg src={image} />
+      )}
+    </AvatarContainer>
   );
-};
-Avatar.propTypes = {
-  onChange: PropTypes.func,
 };
 
 export default Avatar;
